+++ .prettierignore
@@ -1,29 +1,13 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
 # Dependencies
-/node_modules
-/.pnp
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/versions
-
-# Testing
-/coverage
-/.nyc_output
-/junit.xml
-
-# Next.js
-/.next/
-/out/
-/build
-/dist
+node_modules/
+.pnp
+.pnp.js
 
 # Production builds
-/build
-/dist
+.next/
+out/
+build/
+dist/
 
 # Environment variables
 .env
@@ -32,28 +16,10 @@
 .env.test.local
 .env.production.local
 
-# Vercel
-.vercel
-
-# TypeScript
-*.tsbuildinfo
-next-env.d.ts
-
-# Prisma
-/src/generated/prisma
-/prisma/migrations/
-
-# Database
-*.db
-*.sqlite
-*.sqlite3
-
 # Logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.pnpm-debug.log*
-lerna-debug.log*
 
 # Runtime data
 pids
@@ -64,6 +30,9 @@ pids
 # Coverage directory used by tools like istanbul
 coverage/
 *.lcov
+
+# nyc test coverage
+.nyc_output
 
 # Dependency directories
 jspm_packages/
@@ -128,16 +97,13 @@ temp/
 ehthumbs.db
 Thumbs.db
 
+# Prisma
+prisma/migrations/
+
+# Database
+*.db
+*.sqlite
+*.sqlite3
+
 # Generated files
 *.generated.*
-
-# Test results
-test-results/
-playwright-report/
-playwright/.cache/
-
-# Sentry
-.sentryclirc
-
-# Local development
-.local
