+++ README.md
@@ -1,11 +1,12 @@
 # 🚀 C2P Platform - Plateforme d'Apprentissage Développeur
 
-[![Next.js](https://img.shields.io/badge/Next.js-14-black?style=for-the-badge&logo=next.js)](https://nextjs.org/)
+[![Next.js](https://img.shields.io/badge/Next.js-15-black?style=for-the-badge&logo=next.js)](https://nextjs.org/)
 [![TypeScript](https://img.shields.io/badge/TypeScript-5.0-blue?style=for-the-badge&logo=typescript)](https://www.typescriptlang.org/)
 [![Prisma](https://img.shields.io/badge/Prisma-6.15-2D3748?style=for-the-badge&logo=prisma)](https://www.prisma.io/)
 [![Tailwind CSS](https://img.shields.io/badge/Tailwind_CSS-3.3-38B2AC?style=for-the-badge&logo=tailwind-css)](https://tailwindcss.com/)
+[![React](https://img.shields.io/badge/React-19-61DAFB?style=for-the-badge&logo=react)](https://reactjs.org/)
 
-> **C2P Platform** est une plateforme d'apprentissage en ligne moderne et interactive conçue pour les développeurs de tous niveaux. Elle combine formation structurée, communauté active et gamification pour créer une expérience d'apprentissage engageante.
+> **C2P Platform** est une plateforme d'apprentissage en ligne révolutionnaire conçue pour les développeurs de tous niveaux. Elle combine formation structurée, communauté active, gamification et un éditeur de code interactif pour créer une expérience d'apprentissage immersive et engageante.
 
 ## ✨ **Fonctionnalités Principales**
 
@@ -27,25 +28,35 @@
 - **Défis** et objectifs d'apprentissage
 - **Portfolio de projets** public
 
+### 💻 **Éditeur de Code Interactif**
+- **Sandbox HTML/CSS** avec aperçu en temps réel
+- **Objectifs d'apprentissage** progressifs
+- **Auto-complétion** et formatage automatique
+- **Mode focus** pour une concentration optimale
+- **Système de badges** pour les accomplissements
+
 ### 📱 **Interface Moderne**
 - **Design responsive** optimisé mobile/desktop
 - **Thème sombre/clair** automatique
 - **Navigation intuitive** et accessible
-- **Performance optimisée** avec Next.js 14
+- **Performance optimisée** avec Next.js 15
+- **Architecture modulaire** pour une maintenabilité maximale
 
 ## 🛠 **Stack Technique**
 
 | Composant | Technologie | Version |
 |-----------|-------------|---------|
-| **Frontend** | Next.js | 14.x |
+| **Frontend** | Next.js | 15.x |
 | **Language** | TypeScript | 5.x |
+| **React** | React | 19.x |
 | **Styling** | Tailwind CSS | 3.3.x |
 | **Base de données** | PostgreSQL | via Neon |
 | **ORM** | Prisma | 6.15.x |
 | **Authentification** | NextAuth.js | 4.x |
 | **UI Components** | Radix UI + shadcn/ui | Latest |
-| **Gestion d'état** | TanStack Query | 5.x |
+| **Gestion d'état** | TanStack Query + Zustand | 5.x |
 | **Validation** | React Hook Form + Zod | Latest |
+| **Canvas** | HTML5 Canvas API | Native |
 
 ## 🚀 **Installation & Démarrage**
 
@@ -117,7 +128,7 @@ Ouvrez [http://localhost:3000](http://localhost:3000) dans votre navigateur.
 ```
 c2p-platform/
 ├── 📁 src/
-│   ├── 📁 app/                 # App Router Next.js 14
+│   ├── 📁 app/                 # App Router Next.js 15
 │   │   ├── 📁 api/            # Routes API
 │   │   ├── 📁 auth/           # Pages d'authentification
 │   │   ├── 📁 courses/        # Pages des cours
@@ -125,6 +136,22 @@ c2p-platform/
 │   │   ├── 📁 profile/        # Profil utilisateur
 │   │   └── 📁 dashboard/      # Tableau de bord
 │   ├── 📁 components/         # Composants réutilisables
+│   │   ├── 📁 sections/       # Composants de sections
+│   │   │   ├── CapsuleInitiation.tsx
+│   │   │   ├── AnneauCommunaute.tsx
+│   │   │   ├── SandboxLive.tsx
+│   │   │   └── ForgeBadge.tsx
+│   │   ├── 📁 ui/             # Composants UI
+│   │   │   ├── CTAHologramme.tsx
+│   │   │   ├── TypographieDynamique.tsx
+│   │   │   ├── LoadingSpinner.tsx
+│   │   │   └── Skeleton.tsx
+│   │   └── 📁 canvas/         # Composants Canvas
+│   │       ├── PulseActivite.tsx
+│   │       └── CanvasInteractif.tsx
+│   ├── 📁 hooks/              # Hooks personnalisés
+│   │   ├── useEditor.ts       # Hook pour l'éditeur
+│   │   └── useAuth.tsx        # Hook d'authentification
 │   ├── 📁 lib/                # Utilitaires et configurations
 │   └── 📁 types/              # Types TypeScript
 ├── 📁 prisma/
@@ -204,10 +231,15 @@ vercel --prod
 - [x] Gestion des cours et leçons
 - [x] Forum communautaire basique
 - [x] Interface utilisateur responsive
+- [x] Architecture modulaire refactorisée
+- [x] Éditeur de code interactif
+- [x] Système de badges basique
 
 ### **Phase 2 - Fonctionnalités Avancées** 🚧
+- [x] Éditeur HTML/CSS avec aperçu temps réel
+- [x] Objectifs d'apprentissage progressifs
+- [x] Composants Canvas interactifs
 - [ ] Système de badges complet
-- [ ] Gamification avancée
 - [ ] Notifications en temps réel
 - [ ] API mobile
 
@@ -216,6 +248,8 @@ vercel --prod
 - [ ] Système de mentorat
 - [ ] Intégrations tierces
 - [ ] Analytics avancés
+- [ ] Tests unitaires et E2E
+- [ ] PWA (Progressive Web App)
 
 ## 🐛 **Dépannage**
