+++ src/app/api/badges/route.ts
@@ -1,36 +1,32 @@
-import { NextRequest, NextResponse } from 'next/server'
-import { prisma } from '@/lib/prisma'
+import { NextResponse } from 'next/server'
 
-export async function GET(request: NextRequest) {
-  try {
-    const { searchParams } = new URL(request.url)
-    const userId = searchParams.get('userId')
-
-    if (!userId) {
-      return NextResponse.json({ error: 'Missing userId' }, { status: 400 })
-    }
-
-    const userBadges = await prisma.userBadge.findMany({
-      where: { userId },
-      include: { badge: true },
-      orderBy: { earnedAt: 'desc' }
-    })
-
-    const data = userBadges.map(ub => ({
-      id: ub.id,
-      name: ub.badge.name,
-      description: ub.badge.description,
-      rarity: ub.badge.rarity,
-      points: ub.badge.points,
-      imageUrl: ub.badge.imageUrl,
-      earnedAt: ub.earnedAt
-    }))
-
-    return NextResponse.json({ badges: data })
-  } catch (error) {
-    console.error('GET /api/badges error:', error)
-    return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 })
-  }
+export async function GET() {
+  return NextResponse.json({
+    badges: [
+      {
+        id: 1,
+        name: 'Initiation HTML',
+        description: 'Premier pas dans le HTML',
+        icon: '🏷️',
+        rarity: 'common',
+        requirements: 'Compléter le premier objectif HTML'
+      },
+      {
+        id: 2,
+        name: 'Maître CSS',
+        description: 'Expert en styles CSS',
+        icon: '🎨',
+        rarity: 'rare',
+        requirements: 'Compléter 5 objectifs CSS'
+      },
+      {
+        id: 3,
+        name: 'Développeur Full-Stack',
+        description: 'Maîtrise complète du développement',
+        icon: '🚀',
+        rarity: 'legendary',
+        requirements: 'Compléter tous les parcours'
+      }
+    ]
+  })
 }
-
-
