+++ src/components/canvas/CanvasInteractif.tsx
@@ -1,0 +1,127 @@
+'use client'
+
+import { useRef, useState, useEffect } from 'react'
+
+export function CanvasInteractif() {
+  const canvasRef = useRef<HTMLCanvasElement>(null)
+  const [mousePos, setMousePos] = useState({ x: 0, y: 0 })
+  const particlesRef = useRef<Array<{
+    x: number
+    y: number
+    vx: number
+    vy: number
+    size: number
+    opacity: number
+    color: string
+  }>>([])
+
+  useEffect(() => {
+    const canvas = canvasRef.current
+    if (!canvas) return
+
+    const ctx = canvas.getContext('2d')
+    if (!ctx) return
+
+    const resizeCanvas = () => {
+      canvas.width = window.innerWidth
+      canvas.height = window.innerHeight
+    }
+
+    resizeCanvas()
+    window.addEventListener('resize', resizeCanvas)
+
+    // Initialiser les particules (densité/opa adaptées au viewport)
+    const initParticles = () => {
+      particlesRef.current = []
+      const total = window.innerWidth < 640 ? 30 : 60
+      for (let i = 0; i < total; i++) {
+        particlesRef.current.push({
+          x: Math.random() * canvas.width,
+          y: Math.random() * canvas.height,
+          vx: (Math.random() - 0.5) * 0.5,
+          vy: (Math.random() - 0.5) * 0.5,
+          size: Math.random() * 2 + 1,
+          opacity: Math.random() * 0.3 + 0.1,
+          color: `hsl(${Math.random() * 60 + 240}, 70%, 60%)`
+        })
+      }
+    }
+
+    initParticles()
+
+    const animate = () => {
+      ctx.clearRect(0, 0, canvas.width, canvas.height)
+
+      particlesRef.current.forEach((particle, index) => {
+        // Mise à jour position
+        particle.x += particle.vx
+        particle.y += particle.vy
+
+        // Rebond sur les bords
+        if (particle.x < 0 || particle.x > canvas.width) particle.vx *= -1
+        if (particle.y < 0 || particle.y > canvas.height) particle.vy *= -1
+
+        // Attraction vers la souris
+        const dx = mousePos.x - particle.x
+        const dy = mousePos.y - particle.y
+        const distance = Math.sqrt(dx * dx + dy * dy)
+        
+        if (distance < 150) {
+          const force = (150 - distance) / 150
+          particle.vx += (dx / distance) * force * 0.01
+          particle.vy += (dy / distance) * force * 0.01
+        }
+
+        // Dessiner la particule
+        ctx.beginPath()
+        ctx.arc(particle.x, particle.y, particle.size, 0, Math.PI * 2)
+        ctx.fillStyle = particle.color
+        ctx.globalAlpha = particle.opacity
+        ctx.fill()
+
+        // Lignes de connexion
+        particlesRef.current.slice(index + 1).forEach(otherParticle => {
+          const dx = particle.x - otherParticle.x
+          const dy = particle.y - otherParticle.y
+          const distance = Math.sqrt(dx * dx + dy * dy)
+
+          if (distance < 100) {
+            ctx.beginPath()
+            ctx.moveTo(particle.x, particle.y)
+            ctx.lineTo(otherParticle.x, otherParticle.y)
+            ctx.strokeStyle = particle.color
+            ctx.globalAlpha = (100 - distance) / 100 * 0.2
+            ctx.lineWidth = 0.5
+            ctx.stroke()
+          }
+        })
+      })
+
+      requestAnimationFrame(animate)
+    }
+
+    animate()
+
+    return () => {
+      window.removeEventListener('resize', resizeCanvas)
+    }
+  }, [mousePos])
+
+  const handleMouseMove = (e: React.MouseEvent<HTMLCanvasElement>) => {
+    const rect = canvasRef.current?.getBoundingClientRect()
+    if (rect) {
+      setMousePos({
+        x: e.clientX - rect.left,
+        y: e.clientY - rect.top
+      })
+    }
+  }
+
+  return (
+    <canvas
+      ref={canvasRef}
+      onMouseMove={handleMouseMove}
+      className="absolute inset-0 w-full h-full pointer-events-none opacity-20 dark:opacity-15 z-0"
+    />
+  )
+}
