+++ src/components/canvas/index.ts
@@ -1,0 +1,2 @@
+export { PulseActivite } from './PulseActivite'
+export { CanvasInteractif } from './CanvasInteractif'
