+++ src/components/sections/index.ts
@@ -1,0 +1,4 @@
+export { CapsuleInitiation } from './CapsuleInitiation'
+export { AnneauCommunaute } from './AnneauCommunaute'
+export { SandboxLive } from './SandboxLive'
+export { ForgeBadge } from './ForgeBadge'
