+++ src/components/ui/Skeleton.tsx
@@ -1,0 +1,37 @@
+'use client'
+
+interface SkeletonProps {
+  className?: string
+  children?: React.ReactNode
+}
+
+export function Skeleton({ className = '', children }: SkeletonProps) {
+  return (
+    <div className={`animate-pulse bg-gray-200 dark:bg-gray-700 rounded ${className}`}>
+      {children}
+    </div>
+  )
+}
+
+export function CardSkeleton() {
+  return (
+    <div className="rounded-2xl border border-gray-200 dark:border-white/10 bg-white/80 dark:bg-white/5 backdrop-blur-xl p-6">
+      <Skeleton className="h-6 w-3/4 mb-4" />
+      <Skeleton className="h-4 w-full mb-2" />
+      <Skeleton className="h-4 w-2/3" />
+    </div>
+  )
+}
+
+export function TextSkeleton({ lines = 3 }: { lines?: number }) {
+  return (
+    <div className="space-y-2">
+      {Array.from({ length: lines }).map((_, i) => (
+        <Skeleton 
+          key={i} 
+          className={`h-4 ${i === lines - 1 ? 'w-2/3' : 'w-full'}`} 
+        />
+      ))}
+    </div>
+  )
+}
