+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export { CTAHologramme } from './CTAHologramme'
+export { TypographieDynamique } from './TypographieDynamique'
